+++ pittgoogle/exceptions.py
@@ -11,5 +11,5 @@ class OpenAlertError(Exception):
     """Raised when unable to deserialize a Pub/Sub message payload."""
 
 
-class SchemaNotFoundError(Exception):
+class SchemaError(Exception):
     """Raised when a schema with a given name is not found in the registry."""